+++ app/src/components/model/panels/right/RightTabsHeader.js
@@ -26,7 +26,12 @@ export function RightTabsHeader(props) {
   const threatsMap = mitigations?.threatsMap || {};
 
   const { data: suggestions } = useListSuggestionsQuery(modelId);
-  const threatSuggestions = suggestions?.threatsMap[selectedComponent.id] || [];
+  const threatSuggestions = (
+    suggestions?.threatsMap[selectedComponent.id] || []
+  )
+    // Hack for now to remove repetitive stride suggestions from the list.
+    // Will address this later as we do a major update to how suggestions work.
+    .filter((s) => s.source !== "stride");
   const controlSuggestions =
     suggestions?.controlsMap[selectedComponent.id] || [];
   const nSuggestions =