+++ core/src/data/migrations/23_threats_add_severity.sql
@@ -1,0 +1,4 @@
+ALTER TABLE threats
+ADD severity varchar(20) NULL DEFAULT NULL;
+
+